backend/requirements.txt
# fastapi==0.116.1
# uvicorn[standard]==0.35.0
# python-multipart==0.0.20
# sqlalchemy==2.0.43
# pydantic-settings==2.10.1
# faiss-cpu==1.12.0
# numpy==2.1.3
# sentence-transformers==5.1.0
# langdetect==1.0.9
# pypdf==5.9.0
# python-docx==1.2.0
# openpyxl==3.1.5
# python-pptx==1.0.2
# httpx==0.28.1
# pdf2image
# pytesseract 
# pillow

fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.43
pydantic-settings==2.10.1
faiss-cpu==1.12.0
numpy==2.1.3
langdetect==1.0.9
pypdf==5.9.0
python-docx==1.2.0
openpyxl==3.1.5
python-pptx==1.0.2
httpx==0.28.1
requests==2.32.3
pdf2image==1.17.0
pytesseract==0.3.13
pillow==11.0.0